+++ src/components/Accordion/Accordion.tsx
@@ -12,6 +12,7 @@ interface AccordionItem {
 
 interface AccordionProps {
   bordered?: boolean
+  multiselectable?: boolean
   items: AccordionItem[]
   className?: string
 }
@@ -51,7 +52,7 @@ export const AccordionItem = (props: AccordionItem): React.ReactElement => {
 }
 
 export const Accordion = (props: AccordionProps): React.ReactElement => {
-  const { bordered, items, className } = props
+  const { bordered, items, className, multiselectable = false } = props
 
   const [openItems, setOpenState] = useState(
     items.filter((i) => !!i.expanded).map((i) => i.id)
@@ -68,18 +69,26 @@ export const Accordion = (props: AccordionProps): React.ReactElement => {
   const toggleItem = (itemId: AccordionItem['id']): void => {
     const newOpenItems = [...openItems]
     const itemIndex = openItems.indexOf(itemId)
+    const isMultiselectable = multiselectable
 
     if (itemIndex > -1) {
       newOpenItems.splice(itemIndex, 1)
     } else {
-      newOpenItems.push(itemId)
+      if (isMultiselectable) {
+        newOpenItems.push(itemId)
+      } else {
+        newOpenItems.splice(0, newOpenItems.length)
+        newOpenItems.push(itemId)
+      }
     }
-
     setOpenState(newOpenItems)
   }
 
   return (
-    <div className={classes} data-testid="accordion">
+    <div
+      className={classes}
+      data-testid="accordion"
+      aria-multiselectable={multiselectable || undefined}>
       {items.map((item, i) => (
         <AccordionItem
           key={`accordionItem_${i}`}