+++ src/components/forms/ComboBox/ComboBox.tsx
@@ -118,6 +118,18 @@ export const ComboBox = (props: ComboBoxProps): React.ReactElement => {
     ) {
       itemRef.current.focus()
     }
+  }, [state.focusMode, state.focusedOption])
+
+  // If the focused element (activeElement) is outside of the combo box,
+  // make sure the focusMode is BLUR
+  useEffect(() => {
+    if (state.focusMode !== FocusMode.None) {
+      if (!containerRef.current?.contains(window.document.activeElement)) {
+        dispatch({
+          type: ActionTypes.BLUR,
+        })
+      }
+    }
   })
 
   const handleInputKeyDown = (event: KeyboardEvent): void => {
@@ -132,29 +144,51 @@ export const ComboBox = (props: ComboBoxProps): React.ReactElement => {
     } else if (event.key === 'Tab') {
       // Clear button is not visible in this case so manually handle focus
       if (state.isOpen && !state.selectedOption) {
-        event.preventDefault()
+        // If there are filtered options, prevent default
+        // If there are "No Results Found", tab over to prevent a keyboard trap
+        if (state.filteredOptions.length > 0) {
+          event.preventDefault()
+          dispatch({
+            type: ActionTypes.FOCUS_OPTION,
+            option: state.filteredOptions[0],
+          })
+        } else {
+          dispatch({
+            type: ActionTypes.BLUR,
+          })
+        }
+      }
+
+      if (!state.isOpen && state.selectedOption) {
         dispatch({
-          type: ActionTypes.FOCUS_OPTION,
-          option: state.filteredOptions[0],
+          type: ActionTypes.BLUR,
         })
       }
-    } else if (event.key === 'Enter' && state.inputValue !== '') {
+    } else if (event.key === 'Enter' && !state.selectedOption) {
       event.preventDefault()
-      dispatch({ type: ActionTypes.CLOSE_LIST })
+      const selectedOption = state.filteredOptions.find(
+        (option) =>
+          option.label.toLowerCase() === state.inputValue.toLowerCase()
+      )
+      if (selectedOption) {
+        dispatch({
+          type: ActionTypes.SELECT_OPTION,
+          option: selectedOption,
+        })
+      } else {
+        dispatch({ type: ActionTypes.CLEAR })
+      }
     }
   }
 
   const handleInputBlur = (event: FocusEvent<HTMLInputElement>): void => {
-    const { target: elementLosingFocus, relatedTarget: newTarget } = event
+    const { relatedTarget: newTarget } = event
     const newTargetIsOutside =
       !newTarget ||
-      (newTarget instanceof Node &&
-        !containerRef.current?.contains(elementLosingFocus))
+      (newTarget instanceof Node && !containerRef.current?.contains(newTarget))
 
-    if (state.selectedOption?.value) {
-      if (newTargetIsOutside) dispatch({ type: ActionTypes.CLOSE_LIST })
-    } else if (newTargetIsOutside) {
-      dispatch({ type: ActionTypes.CLEAR })
+    if (newTargetIsOutside) {
+      dispatch({ type: ActionTypes.BLUR })
     }
   }
 
@@ -203,7 +237,7 @@ export const ComboBox = (props: ComboBoxProps): React.ReactElement => {
       !newTarget ||
       (newTarget instanceof Node && !containerRef.current?.contains(newTarget))
     ) {
-      dispatch({ type: ActionTypes.CLOSE_LIST })
+      dispatch({ type: ActionTypes.BLUR })
     }
   }